Switch to v2 interfaces
