--- src/main/java/io/aiven/flink/connectors/bigquery/sink/DeliveryGuarantee.java
@@ -1,6 +1,0 @@
-package io.aiven.flink.connectors.bigquery.sink;
-
-public enum DeliveryGuarantee {
-  AT_LEAST_ONCE,
-  EXACTLY_ONCE;
-}
