+++ src/test/java/io/aiven/flink/connectors/bigquery/sink/BigQuerySinkTest.java
@@ -13,6 +13,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
 import java.util.stream.Stream;
+import org.apache.flink.connector.base.DeliveryGuarantee;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.table.api.DataTypes;
 import org.apache.flink.table.api.bridge.java.StreamTableEnvironment;
@@ -53,7 +54,7 @@ public class BigQuerySinkTest {
               true,
               dg,
               CREDENTIALS);
-      var table = BigQuerySink.ensureTableExists(fieldNames, fieldTypes, options);
+      var table = BigQueryDynamicTableSink.ensureTableExists(fieldNames, fieldTypes, options);
       table.delete();
     }
   }
@@ -145,7 +146,7 @@ public class BigQuerySinkTest {
             true,
             DeliveryGuarantee.EXACTLY_ONCE,
             CREDENTIALS);
-    var table = BigQuerySink.ensureTableExists(bqColumnNames, bqColumnTypes, options);
+    var table = BigQueryDynamicTableSink.ensureTableExists(bqColumnNames, bqColumnTypes, options);
     List<Column> columns = new ArrayList<>();
     StringBuilder sb = new StringBuilder();
     for (int i = 0; i < flinkColumnNames.length; i++) {
@@ -298,7 +299,7 @@ public class BigQuerySinkTest {
             true,
             DeliveryGuarantee.EXACTLY_ONCE,
             CREDENTIALS);
-    var table = BigQuerySink.ensureTableExists(bqColumnNames, bqColumnTypes, options);
+    var table = BigQueryDynamicTableSink.ensureTableExists(bqColumnNames, bqColumnTypes, options);
     List<Column> columns = new ArrayList<>();
     StringBuilder sb = new StringBuilder();
     for (int i = 0; i < flinkColumnNames.length; i++) {
